finish UI Setup
