+++ src/style.scss
@@ -6,33 +6,37 @@ body {
   background-color: #f9f9fa
 }
 
-.card-weather {
-  margin-left: auto;
-  margin-right: auto;
-  min-width: 500px;
-  max-width: 600px;
-  background: #e1ecff;
-  background-image: linear-gradient(to left bottom, #d6eef6, #dff0fa, #e7f3fc, #eff6fe, #f6f9ff)
-}
-
-
-.card-weather .card-body:first-child {
-  background: url(https://res.cloudinary.com/dxfq3iotg/image/upload/v1557323760/weather.svg) no-repeat center;
-  background-size: cover
-}
-
 h1,
 h2,
 h3,
 h4,
 h5,
 h6 {
-  font-family: "Poppins", sans-serif;
-  font-weight: 500
+  font-family: "Poppins",
+    sans-serif;
+  font-weight: 600
 }
 
 p {
   font-size: 16px
+}
+
+.input-group {
+  width: 440px;
+}
+
+.card-weather {
+  margin-left: auto;
+  margin-right: auto;
+  min-width: 480px;
+  max-width: 620px;
+  background: #e1ecff;
+  background-image: linear-gradient(to left bottom, #d6eef6, #dff0fa, #e7f3fc, #eff6fe, #f6f9ff)
+}
+
+.card-weather .card-body:first-child {
+  background: url(https://res.cloudinary.com/dxfq3iotg/image/upload/v1557323760/weather.svg) no-repeat center;
+  background-size: cover
 }
 
 .text-gray,
@@ -41,23 +45,22 @@ p {
   color: #969696
 }
 
-
 .card-weather .card-body {
   background: #ffffff
 }
 
-.card-weather .weakly-weather {
+.card-weather .weekly-weather {
   background: #ffffff;
   overflow-x: auto
 }
 
-.card-weather .weakly-weather .weakly-weather-item {
+.card-weather .weekly-weather .weekly-weather-item {
   flex: 0 0 14.28%;
   border-right: 1px solid #f2f2f2;
   padding: 1rem;
   text-align: center
 }
 
-.card-weather .weakly-weather .weakly-weather-item i {
+.card-weather .weekly-weather .weekly-weather-item i {
   font-size: 1.4rem
 }
