+++ src/ui.js
@@ -1,0 +1,13 @@
+const UI = (() => {
+  const init = () => {
+
+  }
+
+  return {
+    init,
+
+  }
+
+})();
+
+export default UI;
